cast secret vals to SecretStrs
