+++ tests/test_config.py
@@ -19,3 +19,8 @@ def test_get_config(api_key):
     github_app = get_configurable(configurable=GitHubApp)
     for attr in ["app_name", "id", "private_key", "webhook_secret"]:
         assert hasattr(github_app, attr)
+
+    secret_vals = [fastapi.key, github_app.private_key, github_app.webhook_secret]
+    for sv in secret_vals:
+        assert repr(sv) == "*****"
+        assert str(sv) == "*****"
